pyproject.toml
[project]
name = "otus-homework-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "django-rq>=3.0.1",
    "djangorestframework>=3.16.0",
    "markdown>=3.8.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "ruff>=0.12.1",
]
